requirements.txt
streamlit
streamlit-folium
shapely
folium
googlemaps
polyline
geopy
pandas
requests
sqlalchemy
psycopg2-binary
Pillow
